Add `HasLoggingServer` instance for `NoContentVerb`
Problem: `HasLoggingServer` instance is defined for servant's Verb.
Servant also has NoContentVerb as a separate data type that can be
used in APIs to denote absence of content. Sadly, there is no
`HasLoggingServer` for it, so it's impossible to use
`serverWithLogging` for APIs with `NoContentVerb`.

Solution: define this instance the same way as for `Verb`.