Export `Positive` type publically
Problem: constructing `PaginationSpec` requires a way to construct
`Positive`, but this is not possible without using methods from `Util`
module which is internal.

The original reason for making this module internal was that many
libraries and projects may define their custom `Positive` type, and I
wanted to prevent the users on relying on API of this type since we
have our own purposes on it.

However, users may need to construct values of `Positive` type
themselves and/or define some instances, so it must be exported
publically.

Solution: add public `Util` module, re-export `Positive` type and
constructors there.

Also groom the methods around `Positive` type.