Add base entries to sorting
Problem: often, even if no ordering is specified, it is highly desired
to return everything in _some_ order:
1. We want results to be ordered deterministically, but e.g. Postgres
does not guarantee this.
2. Pagination stops making sence when no any ordering is guaranteed.

Till now I treated this as client's issue, and that's why examples
recommended to allow sorting on `id`. But we better handle this at
backend. If we go this way, we should encourage always adding the base
sorting.

Solution: add another type-level list to `SortingParams` that stands for
base sorting. So the overall sorting order is lexicographic on
`provided ++ base`, where `provided` is sorting spec specified by the
client, and `base` is what is specified on backend.