Add tests on sorting
Problem: at this point we lack tests on filtering (at all).

Solution: let's add tests on a dummy implementation for now.

Dummy implementation and beam-pg implementation have very few custom
logic (that we didn't touch in this PR), most logic is shared, so this
is sort of makes sense for now.