Do not use `autoexporter`
Problem: at this moment, we want the package to be located on Hackage,
but it still fails to render documentation for packages using any build
tools, see the
[issue](https://github.com/haskell/hackage-server/issues/821).

Moreover, my HLS fails to build the project because of autoexporter as
well.

Solution: replace uses of `autoexporter` with manually written module
lists.