Allow custom logging for endpoints
Problem: sometimes it is convenient to fully disable logs for one or
several servant endpoints, or maybe change the severity level.

Solution: add dedicated combinators.

Now `LoggingApiRec`, that applies logging to the server, carries not
only config but also context that accounts for log modifying
combinators.

Tests on logging are also added, on the general functionality and on the
new functionality.