Get rid of `text-format` dep
Problem: we rely on `text-format` package that is not actively
maintained and does not support the most recent GHC.

Moreover, we effectively use two competing formatting libraries now -
`fmt` and `text-format`. Maybe previously the former had a dependency on
the latter to use the `Buildable` typeclass, but now this dependency
does not exist and `fmt` defines its own `Buildable`.

Solution: strip the `text-format` dependency, import `build` function
from `fmt` instead (that was the only thing we used from `text-format`).